Reduce complexity of distribution and installation of media files
Make doesn't support files with spaces in the names, which is a problem
since there are spaces in the filenames of some of the media files.
Right now, we're building temporary tarballs to workaround the issue,
but it's pretty complex and slow.

This commit takes a different tactic to avoid this issue with make. By
building our own customized rules for installing and uninstalling, we
can drive the shell directly, which will respect the spaces when used
with normal globbing. Two make functions are defined to reduce
duplication, but the important part is that we're not using make
variables or using the automake defined rules that scrape shell stdout
everywhere.

The other trick is to distribute the whole source directory so that
automake doesn't have to build a file list. This does mean that any
unintended files in the directory will be distributed, so care should be
taken before running dist.

[endlessm/eos-shell#1986]