Remove spawn of libreoffice upon session start
Instead of using libreoffice headless to convert the XLS to CSV
at run-time, we'll take on the extra maintenance effort to provide
the XLS file for user display and the pre-generated CSV file
for the rename script.  The CSV file is a hidden file so as not to
confuse the user (and decrease the likelihood of user modification).

[endlessm/eos-shell#3473]