Fix removal of old Samples links
Instead of using a .samples link to keep track of the previous
locale's link, which is a bit frail and was buggy due to the
full path changing when the XDG directory name changes, we just
remove any existing links for any locales that are not current.

Note that any existing .samples links will remain,
but they are hidden links that can safely be ignored.

[endlessm/eos-shell#4698]