Revert "Stop shipping the sample media."
This reverts commit a5f1d50bb55c22d438db76907fdbd4c843709d66.

The packaging change corresponding to this wasn't backported to eos3.8
so the OBS build broke which is blocking 3.8.5~beta2. Revert this change
which wasn't supposed to happen anyway (we're in code freeze). The
xdg-user-dirs-gtk change to remove the call to this script wasn't
backported either. We'll have to decide before releasing 3.8.5 whether
to remove the sample media support from the image builder.

https://phabricator.endlessm.com/T30533